requirements.txt
python-dotenv
pinecone-client
langchain
langchain-google-genai
langchain-pinecone
spacy
